Removes the FieldParser plugin which will match on certin queries and override the search against content (while returning invalid binary data)
